requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
python-dotenv==1.0.0
aiofiles==23.2.1
gunicorn==21.2.0
flask-cors==4.0.0
werkzeug==3.0.1
